Extract fetch utils to separate files
